requirements.txt
albumentations==1.3.0
matplotlib==3.5.3
numpy==1.25.2
opencv_python==4.8.0.76
pandas==1.4.4
Pillow==9.3.0
Pillow==10.1.0
torch==2.0.1
tqdm==4.64.1
